requirements.txt
torch==2.9.1
torchvision==0.24.1
transformers==4.57.5
Pillow==12.1.0
numpy==2.2.6
PyYAML==6.0.3
tqdm==4.67.1
torchmetrics==1.8.2
matplotlib == 3.10.8
pytest == 9.0.2
mypy == 1.19.1
streamlit == 1.54.0